Rename dotfiles to remove initial period
Since I symlink all the dotfiles into various directories anyway,
there's no reason their names have to begin with a dot here. Removing
the dot makes it easier to list files in the repo and looks nicer, so
this commit loses the period. I also changed the name of the tmux 1.3
config file to be nicer.