[zsh] Re-implement 'tm' alias as function
Remove the bin/start-tmux.sh script and reimplement it as a function
'tm'. This repo is meant to be configuration files, and thus not on the
executable $PATH, which means that I had to hard-code the path to
start-tmux.sh into the alias definition; since this broke when I moved
'dotfiles' into the .config directory, I just reimplemented it as a
function -- it's much nicer this way anyhow.

After removing start-tmux.sh it made more sense for the 'bin' directory
to be named 'scripts'; for now the scripts directory is for scripts
which may be occasionally needed to perform some system configuration
task, but are not used regularly and so they don't need to be on the
$PATH.

It may make sense to make the tmux functionality a plugin for oh-my-zsh
eventually if it grows some more.